+++ src/app/app-router/dashboard/page.tsx
@@ -1,5 +1,5 @@
 import { Metadata } from 'next';
-import { WrapLink } from '@/components/WrapLink';
+import { DashboardScreen } from '@/screens/DashboardScreen';
 
 // metaデータの上書き
 // only server component
@@ -8,10 +8,5 @@ export const metadata: Metadata = {
 };
 
 export default function DashboardPage() {
-  return (
-    <>
-      <h1>ダッシュボードです</h1>
-      <WrapLink href='/app-router/job_offers'>求人一覧ページへ</WrapLink>
-    </>
-  );
+  return <DashboardScreen />;
 }
