+++ src/app/app-router/layout.tsx
@@ -1,5 +1,5 @@
-import AxiosProvider from './AxiosProvider';
-import TanstackQueryProvider from './TanstackQueryProvider';
+import AxiosProvider from '@/providers/AxiosProvider';
+import TanstackQueryProvider from '@/providers/TanstackQueryProvider';
 
 export default function AppRouterLayout({ children }: { children: React.ReactNode }) {
   return (
