+++ src/app/app-router/page.tsx
@@ -1,5 +1,5 @@
 import { Metadata } from 'next';
-import { WrapLink } from '@/components/WrapLink';
+import { AppRouterScreen } from '@/screens/AppRouterScreen';
 
 // metaデータの上書き
 // only server component
@@ -8,11 +8,6 @@ export const metadata: Metadata = {
   metadataBase: new URL('https://acme.com'),
 };
 
-export default function Page() {
-  return (
-    <>
-      <h1>App Routerで動くページ</h1>
-      <WrapLink href='/app-router/dashboard'>Dashboardページへ</WrapLink>
-    </>
-  );
+export default function AppRouterPage() {
+  return <AppRouterScreen />;
 }
