+++ src/app/app-router/tanstack/page.tsx
@@ -1,4 +1,4 @@
-import { TanstackScreen } from '../../../screens/app-router/tanstack/index';
+import { TanstackScreen } from '@/screens/TanstackScreen/index';
 
 // NOTE: Axiosをかました時にレスポンスの返却値が変わるので、同じようにwrapしたい
 // import { getGetPostsMock } from '@/api/backend.msw'
