+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
-import './globals.css';
-import Header from '../components/Header/index';
+import '@/styles/globals.css';
+import Header from '../components/shared/Header/index';
 
 const inter = Inter({ subsets: ['latin'] });
