+++ src/screens/AppRouterScreen/index.tsx
@@ -1,0 +1,10 @@
+import { WrapLink } from '@/components/shared/WrapLink';
+
+export const AppRouterScreen = () => {
+  return (
+    <>
+      <h1>App Routerで動くページ</h1>
+      <WrapLink href='/app-router/dashboard'>Dashboardページへ</WrapLink>
+    </>
+  );
+};
