+++ src/screens/JobOfferDetailScreen/index.tsx
@@ -1,0 +1,3 @@
+export const JobOfferDetailScreen = () => {
+  return <h1>求人詳細です</h1>;
+};
